A eSet/MSnSet object with grouped sample or feature variables

This object extends the MSnbase::MSnSet class by adding slots necessary to record the grouping structure. These additional slots are inherited from the dplyr::grouped_df class.

Objects of class `Grouped_eSet` are created by the `group_by`
function. See the vignette for examples.

# S4 method for Grouped_eSet
show(object)

Arguments

object
An `Grouped_eSet` instance.

Slots

fvars

a character vector or a list of grouping feature variables.

fdrop

a logical (default is TRUE) to preserve all feature variable factor levels, even those without data.

findices

indices defining the feature groups.

fgroup_sizes

an integer defining the sizes of the feature groups.

fbiggest_group_size

an integer defining the biggest feature group size.

flabels

a data.frame defining the feature groups and levels.

pvars

a character vector or a list of grouping sample variables.

pdrop

a logical (default is TRUE) to preserve all sample variable factor levels, even those without data.

pindices

indices defining the sample groups.

pgroup_sizes

an integer defining the sizes of the sample groups.

pbiggest_group_size

an integer defining the biggest sample group size.

plabels

a data.frame defining the sample groups and levels.