A Grammar of Data Manipulation for Quantitative Proteomics

The `tidyms` package implements tidy principles as defined in the tidyverse packages to omics-type data classes, with (currently at least), an emphasis on quantitative proteomics data.

# S3 method for eSet
arrange(.data, ...)

# S3 method for eSet
filter(.data, ...)

# S3 method for eSet
group_by(.data, ..., add = FALSE)

# S3 method for eSet
select(.data, ...)

# S3 method for Grouped_eSet
summarise(.data, ...)

# S3 method for eSet
as_tibble(x, ..., fcols = fvarLabels(x))

Arguments

.data
An object of class MSnbase::MSnSet.

...
Expressions evaluated in the context of the object's feature and sample variable and passed to the dplyr functions. Ignored in as_tibble,MSnSet.

add
As in the original `dplyr::group_by`function, when ‘add = FALSE’, the default, `group_by()` will override existing groups. To add to the existing groups, use `add = TRUE`.

x
An object of class MSnSet.

fcols
Feature variables to be added. Default is to add all (i.e. fvarLabels(x)). Use NULL to add none.

Details

The vignette provides additional details and examples.

See the original `dplyr` manual pages for details.

See also